refactor(model-router): extract anthropic-compat-core shared module
No behavior change. anthropic.ts is now a thin delegate to the new
shared core. Hook interface (endpointPath, authHeaders, customHeaders,
promptCache) is declared but unused — wired up in follow-up commits.
